A little closer on weighted entries.
LIRS logic no longer assumes at times that an entry == 1 unit of
capacity. Current add works, but update fails to evict properly.

A separate concern is that non-resident entries are not left in the map
yet. When doing so, this could have a large impact on all atomic map
operations, like remove. Unknown is how trivial/frustrating this change
will be.